# openapi-runtime.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API CRUD Empleados y Departamentos",
    "description": "API para gestión de empleados y departamentos",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
    "/api/v1/empleados/{clave}": {
      "get": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "obtenerEmpleado",
        "parameters": [
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^E-\\d{3}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "actualizarEmpleado",
        "parameters": [
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^E-\\d{3}$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "eliminarEmpleado",
        "parameters": [
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^E-\\d{3}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/departamentos/{id}": {
      "get": {
        "tags": [
          "departamento-controller"
        ],
        "operationId": "obtenerPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DepartamentoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "departamento-controller"
        ],
        "operationId": "actualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DepartamentoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "departamento-controller"
        ],
        "operationId": "eliminar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/empleados": {
      "get": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "listarEmpleados",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "nombre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clave",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^E-\\d{3}$"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "pattern": "^(asc|desc)$"
            }
          },
          {
            "name": "departamentoId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoPageResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "crearEmpleado",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/departamentos": {
      "get": {
        "tags": [
          "departamento-controller"
        ],
        "operationId": "listar",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DepartamentoPageResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "departamento-controller"
        ],
        "operationId": "crear",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentoCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DepartamentoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/departamentos/{id}/empleados": {
      "get": {
        "tags": [
          "departamento-controller"
        ],
        "operationId": "listarEmpleadosPorDepartamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoPageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmpleadoUpdateRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "direccion": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "telefono": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 10
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "departamentoId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "direccion",
          "nombre",
          "telefono",
          "version"
        ]
      },
      "DepartamentoSummaryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
      "EmpleadoResponse": {
        "type": "object",
        "properties": {
          "clave": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "departamento": {
            "$ref": "#/components/schemas/DepartamentoSummaryDTO"
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
      "DepartamentoUpdateRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          }
        },
        "required": [
          "nombre"
        ]
      },
      "DepartamentoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
      "EmpleadoCreateRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "direccion": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "telefono": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 10
          },
          "departamentoId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "direccion",
          "nombre",
          "telefono"
        ]
      },
      "DepartamentoCreateRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          }
        },
        "required": [
          "nombre"
        ]
      },
      "EmpleadoPageResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmpleadoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "DepartamentoPageResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepartamentoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}